Validation of input fields. An error handler for file upload. Error handling and logic for regex deleting.

+++ simple_mvc/src/main/java/org/example/web/controllers/BookShelfController.java
@@ -3,19 +3,24 @@ package org.example.web.controllers;
 import org.apache.log4j.Logger;
 import org.example.app.services.BookService;
 import org.example.web.dto.Book;
+import org.example.web.dto.BookIdToRemove;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import javax.validation.Valid;
+import java.io.*;
+import java.util.regex.PatternSyntaxException;
 
 @Controller
 @RequestMapping(value = "/books")
+@Scope("singleton")
 public class BookShelfController {
 
-    private static final String REDIRECT_TO_BOOKSHELF = "redirect:/books/shelf";
     private final Logger logger = Logger.getLogger(BookShelfController.class);
     private final BookService bookService;
 
@@ -26,36 +31,89 @@ public class BookShelfController {
 
     @GetMapping("/shelf")
     public String books(Model model) {
-        logger.info("got book shelf");
+        logger.info(this.toString());
         model.addAttribute("book", new Book());
+        model.addAttribute("bookIdToRemove", new BookIdToRemove());
         model.addAttribute("bookList", bookService.getAllBooks());
         return "book_shelf";
     }
 
     @PostMapping("/save")
-    public String saveBook(Book book) {
-        if (book.getAuthor() != null && book.getTitle() != null && book.getSize() != null) {
+    public String saveBook(@Valid Book book, BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("book", book);
+            model.addAttribute("bookIdToRemove", new BookIdToRemove());
+            model.addAttribute("bookList", bookService.getAllBooks());
+            return "book_shelf";
+        } else {
             bookService.saveBook(book);
             logger.info("current repository size: " + bookService.getAllBooks().size());
+            return "redirect:/books/shelf";
         }
-        return REDIRECT_TO_BOOKSHELF;
     }
 
     @PostMapping("/remove")
-    public String removeBook(@RequestParam(value = "bookIdToRemove", required = false, defaultValue = "") Integer bookIdToRemove, Model model) {
-        if (!bookService.removeBookById(bookIdToRemove)) {
-            model.addAttribute("errorMessage", "Could not find book with id " + bookIdToRemove);
+    public String removeBook(@Valid BookIdToRemove bookIdToRemove, BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("book", new Book());
+            model.addAttribute("bookList", bookService.getAllBooks());
+            return "book_shelf";
+        } else {
+            bookService.removeBookById(bookIdToRemove.getId());
+            return "redirect:/books/shelf";
         }
-        return REDIRECT_TO_BOOKSHELF;
     }
 
     @PostMapping("/removeByRegex")
-    public String removeBookByRegex(@RequestParam(value = "field", required = false, defaultValue = "") String field,
+    public String removeBookByRegex(@RequestParam(value = "field") String field,
                                     @RequestParam(value = "regex", required = false, defaultValue = "") String regex,
                                     Model model) {
-        if (!bookService.removeItemsByRegex(field, regex)) {
-            model.addAttribute("errorMessage", "Could not find any books with specified parameters.");
+        if (regex.isEmpty()) {
+            model.addAttribute("errorMessage", "Regular expression cannot be empty.");
+        } else {
+            try {
+                if (!bookService.removeItemsByRegex(field, regex)) {
+                    model.addAttribute("errorMessage", "Could not find any books with specified parameters.");
+                }
+            } catch (PatternSyntaxException e) {
+                model.addAttribute("errorMessage", "Invalid regular expression syntax.");
+            }
         }
-        return REDIRECT_TO_BOOKSHELF;
+
+        if (model.containsAttribute("errorMessage")) {
+            model.addAttribute("book", new Book());
+            model.addAttribute("bookIdToRemove", new BookIdToRemove());
+            model.addAttribute("bookList", bookService.getAllBooks());
+            return "book_shelf";
+        }
+
+        return "redirect:/books/shelf";
+    }
+
+    @PostMapping("/uploadFile")
+    public String uploadFile(@RequestParam("file") MultipartFile file) throws IOException {
+        String name = file.getOriginalFilename();
+        byte[] bytes = file.getBytes();
+
+        //create dir
+        String rootPath = System.getProperty("catalina.home");
+        File dir = new File(rootPath + File.separator + "external_uploads");
+        if (!dir.exists()) {
+            dir.mkdirs();
+        }
+
+        //create file
+        File serverFile = new File(dir.getAbsolutePath() + File.separator + name);
+        BufferedOutputStream stream = new BufferedOutputStream(new FileOutputStream(serverFile));
+        stream.write(bytes);
+        stream.close();
+        logger.info("new file saved at: " + serverFile.getAbsolutePath());
+        return "redirect:/books/shelf";
+    }
+
+    @ExceptionHandler({IOException.class})
+    public String handleError(Model model, Exception exception) {
+        model.addAttribute("errorMessage", exception.getMessage());
+        return "errors/500";
     }
 }
